--- src/main/scala/com/github/unchama/seasonalevents/limitedlogin/LoginBonusItemId.scala
@@ -1,11 +1,0 @@
-package com.github.unchama.seasonalevents.limitedlogin
-
-sealed trait LoginBonusItemId
-
-case class LoginBonus(itemId: LoginBonusItemId, amount: Int)
-
-//region 配布するものの列挙
-
-case object LoginBonusGachaTicket extends LoginBonusItemId
-
-//endregion
