+++ src/main/scala/com/github/unchama/seichiassist/subsystems/seasonalevents/limitedlogin/LoginBonusItemId.scala
@@ -1,0 +1,11 @@
+package com.github.unchama.seichiassist.subsystems.seasonalevents.limitedlogin
+
+sealed trait LoginBonusItemId
+
+case class LoginBonus(itemId: LoginBonusItemId, amount: Int)
+
+//region 配布するものの列挙
+
+case object LoginBonusGachaTicket extends LoginBonusItemId
+
+//endregion
