+++ src/main/scala/com/github/unchama/seichiassist/subsystems/seasonalevents/newyear/README.md
